Consolidate pod LLM helpers
